event adaptatsia buttu
