+++ src/App.scss
@@ -2,7 +2,7 @@
 @import "../src/assets/styles/Footer.scss";
 @import "../src/assets/styles/Hero.scss";
 @import "../src//assets/styles/Vacancies.scss";
-@import "../src//assets/styles/Eventes.scss";
+@import "../src/assets/styles/Event.scss";
 @import "../src//assets/styles/Community.scss";
 @import "../src//assets/styles/Organizations.scss";
 @import "../src//assets/styles/Video.scss";
