--- src/assets/styles/Eventes.scss
@@ -1,82 +1,0 @@
-#events {
-    margin: 0;
-    padding: 0;
-}
- 
-.events {
-    .btn {
-        margin-left: 78%;
-        button {
-            width: 200px;
-            height: 40px;
-            background-color: rgb(116, 184, 116);
-            color: white;
-            font-size: 15px;
-            font-weight: bold;
-        }
-    }
-
-    .image {
-        position: relative;
-        width: 100%;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-
-     img {
-        border-radius: 5px;
-        margin: 50px;
-        width: 1200px;
-        height: 400px;
-     }
-  
-     .text {
-        background-color: rgba(0,0,0,.3);
-        position: absolute;
-        backdrop-filter: blur(8px);
-        border-radius: 5px;
-        z-index: 10;
-        color: white;
-        width: 1200px;
-        height: 400px;
-        h2 {
-            position: absolute;
-            top: 30px;
-            left: 40px;
-            font-size: 20px;
-        }
-        h1 {
-            position: absolute;
-            top: 240px;
-            left: 50px;
-          
-        }
-
-        h3 {
-            display: flex;
-            flex-direction: column;
-            position: absolute;
-            top: 300px;
-            left: 50px;
-            font-size: 12px;
-            span {
-                font-size: 20px;
-            }
-        }
-
-        h4 {
-            position: absolute;
-            display: flex;
-            flex-direction: column;
-            top: 300px;
-            left: 250px;
-            font-size: 12px;
-
-            span {
-                font-size: 20px;
-            }
-        }
-
-     }
-    }
-}
