+++ src/assets/styles/Vacancies.scss
@@ -46,7 +46,7 @@
     justify-content: center;
    
   }
-  .images {
+  .images    {
     width: 32px;
     height: 32px;
     border-radius: 2px;
