+++ src/assets/styles/media-styles/Menu.scss
@@ -6,7 +6,7 @@
     position: absolute;
     top: 0;
     right: 0;
-    z-index: 8;
+    z-index: 100;
     &--logo {
         display: flex;
         align-items: center;
